Change the MailQRecord to a custom class instead of a namedtuple.
The MailQRecord has been changed to a custom class. The goal was to be
able to facilitate handling of special methods, like DNS Lookups and
such. That hasn't been implemented yet because I'd like it to be done
without requiring another 3rd party library. Also implemented a "domain"
property to get the domain name of a user. Also implemented a "user"
property.