Added additional header and footer parsing.
When testing on a CentOS 6 box I found that the header and footers for
the mail queue output has changed. So now the header and footer checks
an array of potential footer/header types. At some point will need to
provide an interface to allow for registration of additional patterns.